sillyHacks/requirements.txt
asgiref
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==2.9.2
dj-database-url==0.5.0
Django
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.11.1
drf-jwt==1.17.1
gunicorn==20.0.4
idna==2.10
psycopg2 
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==4.1.4
